pyproject.toml
[tool.poetry]
name = "kindle"
version = "0.1.0"
description = ""
authors = ["Moritz Wolf <code@moritzwolf.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
